+++ chat/SettingsView.vue
@@ -199,13 +199,6 @@
                 </label>
             </div>
 
-            <div class="form-group">
-                <label class="control-label" for="risingDisableWindowsHighContrast">
-                    <input type="checkbox" id="risingDisableWindowsHighContrast" v-model="risingDisableWindowsHighContrast"/>
-                    Disable high-contrast mode (Windows)
-                </label>
-            </div>
-
         </div>
         <div v-show="selectedTab === '3'">
             <template v-if="hidden.length">
@@ -275,7 +268,6 @@
 
         risingShowUnreadOfflineCount!: boolean;
         risingColorblindMode!: boolean;
-        risingDisableWindowsHighContrast!: boolean;
 
 
         async load(): Promise<void> {
@@ -313,7 +305,6 @@
             this.risingShowUnreadOfflineCount = settings.risingShowUnreadOfflineCount;
 
             this.risingColorblindMode = settings.risingColorblindMode;
-            this.risingDisableWindowsHighContrast = settings.risingDisableWindowsHighContrast;
         }
 
         async doImport(): Promise<void> {
@@ -369,7 +360,6 @@
                 risingShowUnreadOfflineCount: this.risingShowUnreadOfflineCount,
 
                 risingColorblindMode: this.risingColorblindMode,
-                risingDisableWindowsHighContrast: this.risingDisableWindowsHighContrast
             };
             if(this.notifications) await core.notifications.requestPermission();
         }