+++ chat/common.ts
@@ -54,7 +54,6 @@ export class Settings implements ISettings {
 
     risingShowUnreadOfflineCount = true;
     risingColorblindMode = false;
-    risingDisableWindowsHighContrast = false;
 }