+++ chat/interfaces.ts
@@ -221,7 +221,6 @@ export namespace Settings {
 
         readonly risingShowUnreadOfflineCount: boolean;
         readonly risingColorblindMode: boolean;
-        readonly risingDisableWindowsHighContrast: boolean;
     }
 }