+++ helpers/dialog.ts
@@ -4,6 +4,7 @@ export class Dialog {
   static confirmDialog(message: string): boolean {
     const result = remote.dialog.showMessageBoxSync({
       message,
+      title: 'F-Chat Rising',
       type: 'question',
       buttons: ['Yes', 'No'],
       defaultId: 1,